# individuals-api/openapi/individuals-api.yml
openapi: 3.0.0
info:
  title: Individuals API
  description: REST API для регистрации, логина, обновления токена и получения информации о пользователе через Keycloak.
  version: 1.0.0
servers:
  - url: http://localhost:8081/v1

paths:
  /auth/registration:
    post:
      summary: Регистрация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
      responses:
        '201':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Пользователь уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      summary: Аутентификация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Неверный логин или пароль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/refresh-token:
    post:
      summary: Обновление токена доступа
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
      responses:
        '200':
          description: Токен успешно обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Недействительный или просроченный refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/me:
    get:
      summary: Получение данных текущего пользователя
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Данные пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoResponse'
        '401':
          description: Недействительный токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /individual/delete:
    delete:
      summary: Удаление (деактивация) аккаунта
      responses:
        '200':
          description: Пользователь удален (деактивирован)
        '401':
          description: Недействительный токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /individual/update:
    put:
      summary: Обновление пользовательских данных
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Успешное обновление данных
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Недействительный токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserRegistrationRequest:
      type: object
      required: [ email, password, confirm_password, passport_number, phone_number, status, email, first_name, last_name, address, zip_code, city, state, name, alpha2, alpha3 ]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        confirm_password:
          type: string
        passport_number:
          type: string
        phone_number:
          type: string
        secret_key:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        address:
          type: string
        zip_code:
          type: string
        city:
          type: string
        state:
          type: string
        name:
          type: string
        alpha2:
          type: string
          maxLength: 2
        alpha3:
          type: string
          maxLength: 3

    UserUpdateRequest:
      type: object
      properties:
        passport_number:
          type: string
        phone_number:
          type: string
        secret_key:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        address:
          type: string
        zip_code:
          type: string
        city:
          type: string
        state:
          type: string
        name:
          type: string
        alpha2:
          type: string
          maxLength: 2
        alpha3:
          type: string
          maxLength: 3

    UserLoginRequest:
      type: object
      required: [ email, password ]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    TokenRefreshRequest:
      type: object
      required: [ refresh_token ]
      properties:
        refresh_token:
          type: string

    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        token_type:
          type: string
          example: Bearer

    UserInfoResponse:
      type: object
      properties:
        email:
          type: string
          format: email
        passport_number:
          type: string
        phone_number:
          type: string
        secret_key:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        address:
          type: string
        zip_code:
          type: string
        city:
          type: string
        state:
          type: string
        name:
          type: string
        alpha2:
          type: string
          maxLength: 2
        alpha3:
          type: string
          maxLength: 3

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        status:
          type: integer